test: Refactored how helpers are setup in tests
