+++ tests/conftest.py
@@ -1,0 +1,5 @@
+import sys
+import os
+
+# Add the helpers directory to the path so we can import the helpers
+sys.path.append(os.path.join(os.path.dirname(__file__), "helpers"))
