+++ tests/sims/test_simulator.py
@@ -1,6 +1,7 @@
 import pytest
 
 from caustics.sims.state_dict import StateDict
+from helpers.sims import extract_tensors, isEquals
 
 
 @pytest.fixture
@@ -9,12 +10,12 @@ def state_dict(simple_common_sim):
 
 
 @pytest.fixture
-def expected_tensors(simple_common_sim, sim_utils):
-    return sim_utils.extract_tensors(simple_common_sim.params)
+def expected_tensors(simple_common_sim):
+    return extract_tensors(simple_common_sim.params)
 
 
 class TestSimulator:
-    def test_state_dict(self, state_dict, expected_tensors, sim_utils):
+    def test_state_dict(self, state_dict, expected_tensors):
         # Check state_dict type and default keys
         assert isinstance(state_dict, StateDict)
 
@@ -27,4 +28,4 @@ class TestSimulator:
         assert "created_time" in state_dict._metadata
 
         # Check params
-        assert sim_utils.isEquals(dict(state_dict), expected_tensors)
+        assert isEquals(dict(state_dict), expected_tensors)
